tls_demo_package/scanner/requirements.txt
# Core dependencies
aiohttp==3.8.4
jinja2==3.1.2
typer==0.10.0

# TLS/SSL
cryptography>=35.0.0
pyOpenSSL==23.2.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.0

# Configuration & Data
pydantic>=1.7,<2.0
pyyaml==6.0

# Web crawler
beautifulsoup4==4.12.0
lxml==4.9.0

# Utilities
requests==2.31.0
